refactor: adopt static vector.yaml and v2 product_logging, removing RoleGroupRef
